requirements.txt
streamlit
pandas
plotly
streamlit-option-menu
numerize
mysql-connector-python
python-dotenv